% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetVP-package.R
\docType{package}
\name{UpSetVP-package}
\alias{UpSetVP}
\alias{UpSetVP-package}
\title{Visualization of Variation Partitioning and Hierarchical Partitioning}
\description{
UpSetVP using UpSet matrix layout to visualize the unique, common and individual contribution of each predictor (or matrix) towards explained variation on canonical analysis.
}
\references{
Lai J., Zou Y., Zhang J., Peres-Neto P. (2022) Generalizing hierarchical and variation partitioning in multiple regression and canonical analyses using the rdacca.hp R package. Methods in Ecology and Evolution, 13:782-788.

Liu Y., Yu X., Yu Y., et al. (2023) Application of "rdacca. hp" R package in ecological data analysis: case and progress. Chinese Journal of Plant Ecology, 27: 134-144.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LiuXYh/UpSetVP}
  \item Report bugs at \url{https://github.com/LiuXYh/UpSetVP/issues}
}

}
\keyword{internal}
