% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetVP-package.R
\docType{package}
\name{UpSetVP-package}
\alias{UpSetVP}
\alias{UpSetVP-package}
\title{Visualization of Variation Partitioning and Hierarchical Partitioning}
\description{
UpSetVP using UpSet matrix layout to visualize the unique, common and individual contributions of each predictor (or matrix) towards explained variation on canonical analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LiuXYh/UpSetVP}
}

}
\keyword{internal}
